requirements-prod.txt
# Production dependencies with pinned versions for stability

# Core dependencies
anthropic==0.40.0
click==8.1.7
rich==13.7.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
jinja2==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0

# WebSocket dependencies
python-socketio==5.10.0

# HTTP client
aiohttp==3.9.3
httpx==0.26.0

# File operations
aiofiles==23.2.1

# Environment management
python-dotenv==1.0.0

# Development dependencies (for production debugging)
pytest==7.4.3
pytest-asyncio==0.21.1

# Note: These are the minimum required dependencies for production
# Additional development dependencies are in requirements.txt